remove via tag for signalement persist

+++ src/EventSubscriber/CacheInvalidationSubscriber.php
@@ -12,14 +12,14 @@ use Doctrine\Persistence\Event\LifecycleEventArgs;
 use Psr\Cache\InvalidArgumentException;
 use Psr\Log\LoggerInterface;
 use Symfony\Bundle\SecurityBundle\Security;
-use Symfony\Contracts\Cache\CacheInterface;
+use Symfony\Contracts\Cache\TagAwareCacheInterface;
 
 class CacheInvalidationSubscriber implements EventSubscriberInterface
 {
     public const CONTEXT_WIDGET_DATA_KPI = 'countDataKpi';
 
     public function __construct(
-        readonly private CacheInterface $dashboardCache,
+        readonly private TagAwareCacheInterface $dashboardCache,
         readonly private CacheCommonKeyGenerator $cacheCommonKeyGenerator,
         readonly private LoggerInterface $logger,
         private readonly Security $security,
@@ -62,11 +62,17 @@ class CacheInvalidationSubscriber implements EventSubscriberInterface
             /** @var User $user */
             $user = $this->security->getUser();
             $territory = $user?->getTerritory();
-            $commonKey = $this->cacheCommonKeyGenerator->generate();
-            $key = self::CONTEXT_WIDGET_DATA_KPI.'-'.$commonKey.'-zip-'.$territory?->getZip().'-id-'.$user?->getId();
-
             try {
-                $this->dashboardCache->delete($key);
+                if ($entity instanceof Signalement) {
+                    $this->dashboardCache->invalidateTags(['data-kpi-'.$territory?->getZip()]);
+                } else {
+                    $commonKey = $this->cacheCommonKeyGenerator->generate();
+                    $key = self::CONTEXT_WIDGET_DATA_KPI
+                        .'-'.$commonKey
+                        .'-zip-'.$territory?->getZip()
+                        .'-id-'.$user?->getId();
+                    $this->dashboardCache->delete($key);
+                }
             } catch (InvalidArgumentException $exception) {
                 $this->logger->error(sprintf('Invalidate cache failed %s', $exception->getMessage()));
             }
